# apps/openrouter-facade/contracts/openapi.golden.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pippit Bridge",
    "version": "0.1.0",
    "description": "An OpenRouter-compatible image and asynchronous video API backed by Pippit. BYOK management follows OpenRouter's management-key boundary; provider=pippit is an explicit facade extension."
  },
...
  "paths": {
...
    "/api/v1/videos/{jobId}": {
      "get": {
        "operationId": "getVideo",
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoGenerationJob"
                }
              }
            },
            "description": "Current video generation status"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "502": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "runtimeBearer": []
          }
        ],
        "summary": "Poll video generation status",
        "tags": [
          "Videos"
        ]
      }
    },
    "/api/v1/videos/{jobId}/content": {
      "get": {
        "operationId": "getVideoContent",
        "parameters": [
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "query",
            "name": "index",
            "required": false,
            "schema": {
              "default": 0,
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "video/mp4": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Generated video bytes"
          },
          "206": {
            "content": {
              "video/mp4": {
                "schema": {
                  "format": "binary",
                  "type": "string"
                }
              }
            },
            "description": "Generated video byte range"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "416": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          },
          "502": {
            "$ref": "#/components/responses/Error"
          },
          "504": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "runtimeBearer": []
          }
        ],
        "summary": "Download generated video content",
        "tags": [
          "Videos"
        ]
      }
    },
...
  },
  "components": {
...
    "responses": {
      "Error": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "OpenRouter-style error response"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              },
              "metadata": {
                "additionalProperties": true,
                "type": "object"
              },
              "param": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "type": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "param",
              "type"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "VideoGenerationJob": {
        "properties": {
          "error": {
            "type": "string"
          },
          "generation_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "polling_url": {
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "in_progress",
              "completed",
              "failed",
              "cancelled",
              "expired"
            ],
            "type": "string"
          },
          "unsigned_urls": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "usage": {
            "properties": {
              "cost": {
                "type": [
                  "number",
                  "null"
                ]
              },
              "is_byok": {
                "type": "boolean"
              }
            },
            "type": "object"
          }
        },
        "required": [
          "id",
          "polling_url",
          "status"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "managementBearer": {
        "bearerFormat": "Management API Key",
        "description": "Server management key used only for /api/v1/byok. It is independent from both facade API keys and Pippit AKs.",
        "scheme": "bearer",
        "type": "http"
      },
      "runtimeBearer": {
        "bearerFormat": "Facade API Key",
        "description": "Facade API key used for model discovery and video requests. It is not a Pippit AK; the server resolves an eligible stored BYOK credential.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  }
}